compatible with Redmine 2.1
	use jquery(in Redmine) instead of prototype.js
