+++ app/views/projects/index.html.erb
@@ -30,7 +30,7 @@
       <% project_id = "%04d" % project.id %>
       <%  (plevel - level).times { ancestors.pop } if (level < plevel) %>
       <% has_children = !project.children.empty? %>
-      <tr id="<%= project_id %>" class="<%= project.css_classes %> collapsed <%= level > 0 ? "hide #{ancestors.last}" : cycle("odd", "even") %>">
+      <tr id="<%= project_id %>" class="<%= project.css_classes %> <%= level > 0 ? "hide #{ancestors.last}" : cycle("odd", "even") %>">
         <td class="name" style="padding-left: <%= (18 * level).to_s %>px">
           <% if has_children %>
             <span class="expander" onclick="toggleShowHide(<%= "'#{project_id}'" %>)">&nbsp;</span>
